moved popup into separate file
